Implemented most of the functionality
Will be rewritten shortly, but I wanted to preserve this as a separate commit since it contains valuable code. The Windows Service stuff will be rehauled; the Win32 API GetLastInputInfo() does not return sensible values when being called from a Windows Service.

I _could_ make that part a systray app that runs and communicates with a Windows Service using local RPC, but it feels completely overkill. It's better to just let this be a regular in-the-background Windows app instead (without a window) - KISS.